Random graph generation
A class for random graph generation is added. The algorithm is based on
the first algorithm find in the article Fast Random Graph Generation,
National University of Singapore with constraints for self vertex loops
and parallel edges.

Second class for time bench marking of the Mincost Flow algorithm with
the random generated graph is added.